+++ src/resources/meta.ts
@@ -1,0 +1,22 @@
+import type { CsfloatHttpClient } from "../client.js";
+import type {
+  CsfloatExchangeRatesResponse,
+  CsfloatLocationResponse,
+  CsfloatSchemaResponse,
+} from "../types.js";
+
+export class MetaResource {
+  constructor(private readonly client: CsfloatHttpClient) {}
+
+  getSchema(): Promise<CsfloatSchemaResponse> {
+    return this.client.get<CsfloatSchemaResponse>("schema");
+  }
+
+  getExchangeRates(): Promise<CsfloatExchangeRatesResponse> {
+    return this.client.get<CsfloatExchangeRatesResponse>("meta/exchange-rates");
+  }
+
+  getLocation(): Promise<CsfloatLocationResponse> {
+    return this.client.get<CsfloatLocationResponse>("meta/location");
+  }
+}