+++ src/sdk.ts
@@ -3,6 +3,7 @@ import { AccountResource } from "./resources/account.js";
 import { HistoryResource } from "./resources/history.js";
 import { InventoryResource } from "./resources/inventory.js";
 import { ListingsResource } from "./resources/listings.js";
+import { MetaResource } from "./resources/meta.js";
 import { StallResource } from "./resources/stall.js";
 import { UsersResource } from "./resources/users.js";
 
@@ -14,6 +15,7 @@ export class CsfloatSdk {
   readonly listings: ListingsResource;
   readonly history: HistoryResource;
   readonly users: UsersResource;
+  readonly meta: MetaResource;
 
   constructor(options: CsfloatClientOptions) {
     this.client = new CsfloatHttpClient(options);
@@ -23,5 +25,6 @@ export class CsfloatSdk {
     this.listings = new ListingsResource(this.client);
     this.history = new HistoryResource(this.client);
     this.users = new UsersResource(this.client);
+    this.meta = new MetaResource(this.client);
   }
 }